Move all bottom sheet animations from theme to launchers
