+++ sdk/src/main/kotlin/com/processout/sdk/ui/nativeapm/PONativeAlternativePaymentMethodLauncher.kt
@@ -1,15 +1,19 @@
 package com.processout.sdk.ui.nativeapm
 
+import android.content.Context
 import androidx.activity.ComponentActivity
 import androidx.activity.result.ActivityResultLauncher
+import androidx.core.app.ActivityOptionsCompat
 import androidx.fragment.app.Fragment
+import com.processout.sdk.R
 
 /**
  * Launcher that starts [PONativeAlternativePaymentMethodActivity] and provides the result.
  */
-class PONativeAlternativePaymentMethodLauncher private constructor() {
-
-    private lateinit var launcher: ActivityResultLauncher<PONativeAlternativePaymentMethodConfiguration>
+class PONativeAlternativePaymentMethodLauncher private constructor(
+    private val launcher: ActivityResultLauncher<PONativeAlternativePaymentMethodConfiguration>,
+    private val activityOptions: ActivityOptionsCompat
+) {
 
     companion object {
         /**
@@ -19,12 +23,13 @@ class PONativeAlternativePaymentMethodLauncher private constructor() {
         fun create(
             from: Fragment,
             callback: PONativeAlternativePaymentMethodResultCallback
-        ) = PONativeAlternativePaymentMethodLauncher().apply {
+        ) = PONativeAlternativePaymentMethodLauncher(
             launcher = from.registerForActivityResult(
                 PONativeAlternativePaymentMethodActivityContract(),
                 callback::onNativeAlternativePaymentMethodResult
-            )
-        }
+            ),
+            activityOptions = createActivityOptions(from.requireContext())
+        )
 
         /**
          * Creates the launcher from Activity.
@@ -33,19 +38,25 @@ class PONativeAlternativePaymentMethodLauncher private constructor() {
         fun create(
             from: ComponentActivity,
             callback: PONativeAlternativePaymentMethodResultCallback
-        ) = PONativeAlternativePaymentMethodLauncher().apply {
+        ) = PONativeAlternativePaymentMethodLauncher(
             launcher = from.registerForActivityResult(
                 PONativeAlternativePaymentMethodActivityContract(),
                 from.activityResultRegistry,
                 callback::onNativeAlternativePaymentMethodResult
+            ),
+            activityOptions = createActivityOptions(from)
+        )
+
+        private fun createActivityOptions(context: Context) =
+            ActivityOptionsCompat.makeCustomAnimation(
+                context, R.anim.po_slide_in_vertical, 0
             )
-        }
     }
 
     /**
      * Launches the activity.
      */
     fun launch(configuration: PONativeAlternativePaymentMethodConfiguration) {
-        launcher.launch(configuration)
+        launcher.launch(configuration, activityOptions)
     }
 }
