+++ sdk/src/main/res/anim/po_slide_in_vertical.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <translate xmlns:android="http://schemas.android.com/apk/res/android"
-    android:duration="300"
+    android:duration="400"
     android:fromYDelta="100%p"
     android:interpolator="@android:interpolator/fast_out_slow_in"
     android:toYDelta="0" />
