+++ sdk/src/main/res/anim/po_slide_out_vertical.xml
