+++ sdk/src/main/res/values/themes.xml
@@ -41,6 +41,7 @@
     <style name="ThemeOverlay.ProcessOut.Base.BottomSheetDialog" parent="ThemeOverlay.Material3.BottomSheetDialog">
         <item name="android:backgroundDimEnabled">true</item>
         <item name="android:windowSoftInputMode">adjustResize</item>
+        <item name="android:windowAnimationStyle">@null</item>
         <item name="colorSurface">@android:color/transparent</item>
         <item name="bottomSheetStyle">@style/Widget.ProcessOut.BottomSheet.Modal</item>
     </style>
