+++ ui/src/main/kotlin/com/processout/sdk/ui/card/tokenization/POCardTokenizationLauncher.kt
@@ -1,17 +1,21 @@
 package com.processout.sdk.ui.card.tokenization
 
+import android.content.Context
 import androidx.activity.ComponentActivity
 import androidx.activity.result.ActivityResultLauncher
+import androidx.core.app.ActivityOptionsCompat
 import androidx.fragment.app.Fragment
+import com.processout.sdk.R
 import com.processout.sdk.api.model.response.POCard
 import com.processout.sdk.core.ProcessOutActivityResult
 
 /**
  * Launcher that starts [CardTokenizationActivity] and provides the result.
  */
-class POCardTokenizationLauncher private constructor() {
-
-    private lateinit var launcher: ActivityResultLauncher<POCardTokenizationConfiguration>
+class POCardTokenizationLauncher private constructor(
+    private val launcher: ActivityResultLauncher<POCardTokenizationConfiguration>,
+    private val activityOptions: ActivityOptionsCompat
+) {
 
     companion object {
         /**
@@ -21,12 +25,13 @@ class POCardTokenizationLauncher private constructor() {
         fun create(
             from: Fragment,
             callback: (ProcessOutActivityResult<POCard>) -> Unit
-        ) = POCardTokenizationLauncher().apply {
+        ) = POCardTokenizationLauncher(
             launcher = from.registerForActivityResult(
                 CardTokenizationActivityContract(),
                 callback
-            )
-        }
+            ),
+            activityOptions = createActivityOptions(from.requireContext())
+        )
 
         /**
          * Creates the launcher from Activity.
@@ -35,19 +40,25 @@ class POCardTokenizationLauncher private constructor() {
         fun create(
             from: ComponentActivity,
             callback: (ProcessOutActivityResult<POCard>) -> Unit
-        ) = POCardTokenizationLauncher().apply {
+        ) = POCardTokenizationLauncher(
             launcher = from.registerForActivityResult(
                 CardTokenizationActivityContract(),
                 from.activityResultRegistry,
                 callback
+            ),
+            activityOptions = createActivityOptions(from)
+        )
+
+        private fun createActivityOptions(context: Context) =
+            ActivityOptionsCompat.makeCustomAnimation(
+                context, R.anim.po_slide_in_vertical, 0
             )
-        }
     }
 
     /**
      * Launches the activity.
      */
     fun launch(configuration: POCardTokenizationConfiguration) {
-        launcher.launch(configuration)
+        launcher.launch(configuration, activityOptions)
     }
 }
