+++ ui/src/main/kotlin/com/processout/sdk/ui/napm/PONativeAlternativePaymentLauncher.kt
@@ -1,17 +1,21 @@
 package com.processout.sdk.ui.napm
 
+import android.content.Context
 import androidx.activity.ComponentActivity
 import androidx.activity.result.ActivityResultLauncher
+import androidx.core.app.ActivityOptionsCompat
 import androidx.fragment.app.Fragment
+import com.processout.sdk.R
 import com.processout.sdk.core.POUnit
 import com.processout.sdk.core.ProcessOutActivityResult
 
 /**
  * Launcher that starts [NativeAlternativePaymentActivity] and provides the result.
  */
-class PONativeAlternativePaymentLauncher private constructor() {
-
-    private lateinit var launcher: ActivityResultLauncher<PONativeAlternativePaymentConfiguration>
+class PONativeAlternativePaymentLauncher private constructor(
+    private val launcher: ActivityResultLauncher<PONativeAlternativePaymentConfiguration>,
+    private val activityOptions: ActivityOptionsCompat
+) {
 
     companion object {
         /**
@@ -21,12 +25,13 @@ class PONativeAlternativePaymentLauncher private constructor() {
         fun create(
             from: Fragment,
             callback: (ProcessOutActivityResult<POUnit>) -> Unit
-        ) = PONativeAlternativePaymentLauncher().apply {
+        ) = PONativeAlternativePaymentLauncher(
             launcher = from.registerForActivityResult(
                 NativeAlternativePaymentActivityContract(),
                 callback
-            )
-        }
+            ),
+            activityOptions = createActivityOptions(from.requireContext())
+        )
 
         /**
          * Creates the launcher from Activity.
@@ -35,19 +40,25 @@ class PONativeAlternativePaymentLauncher private constructor() {
         fun create(
             from: ComponentActivity,
             callback: (ProcessOutActivityResult<POUnit>) -> Unit
-        ) = PONativeAlternativePaymentLauncher().apply {
+        ) = PONativeAlternativePaymentLauncher(
             launcher = from.registerForActivityResult(
                 NativeAlternativePaymentActivityContract(),
                 from.activityResultRegistry,
                 callback
+            ),
+            activityOptions = createActivityOptions(from)
+        )
+
+        private fun createActivityOptions(context: Context) =
+            ActivityOptionsCompat.makeCustomAnimation(
+                context, R.anim.po_slide_in_vertical, 0
             )
-        }
     }
 
     /**
      * Launches the activity.
      */
     fun launch(configuration: PONativeAlternativePaymentConfiguration) {
-        launcher.launch(configuration)
+        launcher.launch(configuration, activityOptions)
     }
 }
